+++ package.json
@@ -1,6 +1,6 @@
 {
-  "name": "AwesomeProject",
-  "version": "0.1.0",
+  "name": "sample-react-native",
+  "version": "0.1.1",
   "private": true,
   "devDependencies": {
     "react-native-scripts": "1.14.0",
